Relocated important torrent check
The important torrent check was being performed first and never trying to get a good announce quickly.  This change moved the important torrent check into the logic block of the deletion of the torrent.  This way the important torrent will try to be swarmed just like any other, but it if fails then it won't be deleted since it's important.